Allow reattaching in-progress fragments
This change allows for fragments that are still being modified to
replace the current document head and body. Using a MutationObserver, we
wait for the fragment to have the same number of Nodes as the document
and then swap it. This means that if there is a race between
done-autorender and incremental loading, we swap as soon as we can.

Closes #2